fix_services/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
ntlm-auth==1.5.0
paramiko==2.7.2
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
python-decouple==3.4
pywinrm==0.4.1
requests==2.25.1
requests-ntlm==1.1.0
six==1.15.0
urllib3==1.26.4
xmltodict==0.12.0
